Correction de l'export et intégration dans EventAccueil
