+++ WebContent/eventAccueil.xhtml
@@ -251,6 +251,8 @@
 
 						<p:commandButton value="Gérer la grille critériée"
 							actionListener="#{eventAccueil.buttonGrille}" ajax="false" />
+						<p:commandButton value="Exporter les résultats de l'événement"
+							actionListener="#{eventAccueil.buttonResult}" ajax="false" />
 					</div>
 					<br />
 				</div>
